+++ json_serializable/test/kitchen_sink/kitchen_sink.factories.dart
@@ -3,19 +3,13 @@ import 'kitchen_sink.g_any_map.dart' as any_map;
 import 'kitchen_sink.g_any_map__checked__non_nullable.dart'
     as any_map__checked__non_nullable;
 import 'kitchen_sink.g_any_map__non_nullable.dart' as any_map__non_nullable;
-import 'kitchen_sink.g_any_map__non_nullable__use_wrappers.dart'
-    as any_map__non_nullable__use_wrappers;
 import 'kitchen_sink.g_exclude_null.dart' as exclude_null;
 import 'kitchen_sink.g_exclude_null__no_encode_empty.dart'
     as exclude_null__no_encode_empty;
-import 'kitchen_sink.g_exclude_null__no_encode_empty__non_nullable__use_wrappers.dart'
-    as exclude_null__no_encode_empty__non_nullable__use_wrappers;
-import 'kitchen_sink.g_exclude_null__no_encode_empty__use_wrappers.dart'
-    as exclude_null__no_encode_empty__use_wrappers;
+import 'kitchen_sink.g_exclude_null__no_encode_empty__non_nullable.dart'
+    as exclude_null__no_encode_empty__non_nullable;
 import 'kitchen_sink.g_exclude_null__non_nullable.dart'
     as exclude_null__non_nullable;
-import 'kitchen_sink.g_exclude_null__use_wrappers.dart'
-    as exclude_null__use_wrappers;
 import 'kitchen_sink.g_explicit_to_json.dart' as explicit_to_json;
 import 'kitchen_sink.g_no_encode_empty.dart' as no_encode_empty;
 import 'kitchen_sink.g_no_encode_empty__non_nullable.dart'
@@ -26,13 +20,10 @@ const factories = [
   any_map.factory,
   any_map__checked__non_nullable.factory,
   any_map__non_nullable.factory,
-  any_map__non_nullable__use_wrappers.factory,
   exclude_null.factory,
   exclude_null__no_encode_empty.factory,
-  exclude_null__no_encode_empty__non_nullable__use_wrappers.factory,
-  exclude_null__no_encode_empty__use_wrappers.factory,
+  exclude_null__no_encode_empty__non_nullable.factory,
   exclude_null__non_nullable.factory,
-  exclude_null__use_wrappers.factory,
   explicit_to_json.factory,
   no_encode_empty.factory,
   no_encode_empty__non_nullable.factory,