+++ json_serializable/test/src/generic_test_input.dart
@@ -50,56 +50,6 @@ GenericClass<T, S> _$GenericClassFromJson<T extends num, S>(
     ..fieldS = json['fieldS'] == null ? null : _dataFromJson(json['fieldS']);
 }
 
-Map<String, dynamic> _$GenericClassToJson<T extends num, S>(
-        GenericClass<T, S> instance) =>
-    _$GenericClassJsonMapWrapper<T, S>(instance);
-
-class _$GenericClassJsonMapWrapper<T extends num, S> extends $JsonMapWrapper {
-  final GenericClass<T, S> _v;
-  _$GenericClassJsonMapWrapper(this._v);
-
-  @override
-  Iterable<String> get keys =>
-      const ['fieldObject', 'fieldDynamic', 'fieldInt', 'fieldT', 'fieldS'];
-
-  @override
-  dynamic operator [](Object key) {
-    if (key is String) {
-      switch (key) {
-        case 'fieldObject':
-          return _v.fieldObject == null ? null : _dataToJson(_v.fieldObject);
-        case 'fieldDynamic':
-          return _v.fieldDynamic == null ? null : _dataToJson(_v.fieldDynamic);
-        case 'fieldInt':
-          return _v.fieldInt == null ? null : _dataToJson(_v.fieldInt);
-        case 'fieldT':
-          return _v.fieldT == null ? null : _dataToJson(_v.fieldT);
-        case 'fieldS':
-          return _v.fieldS == null ? null : _dataToJson(_v.fieldS);
-      }
-    }
-    return null;
-  }
-}
-''',
-  configurations: ['wrapped'],
-)
-@ShouldGenerate(
-  r'''
-GenericClass<T, S> _$GenericClassFromJson<T extends num, S>(
-    Map<String, dynamic> json) {
-  return GenericClass<T, S>()
-    ..fieldObject =
-        json['fieldObject'] == null ? null : _dataFromJson(json['fieldObject'])
-    ..fieldDynamic = json['fieldDynamic'] == null
-        ? null
-        : _dataFromJson(json['fieldDynamic'])
-    ..fieldInt =
-        json['fieldInt'] == null ? null : _dataFromJson(json['fieldInt'])
-    ..fieldT = json['fieldT'] == null ? null : _dataFromJson(json['fieldT'])
-    ..fieldS = json['fieldS'] == null ? null : _dataFromJson(json['fieldS']);
-}
-
 abstract class _$GenericClassSerializerMixin<T extends num, S> {
   Object get fieldObject;
   dynamic get fieldDynamic;